requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
pytest==7.4.2
pytest-django==4.5.2
